Extract force generators updater to separate class
