--- Assets/Scripts/Features/Services/Transformation/IPositionChanger.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace WizardSpells.Features.Services.Transformation
-{
-    public interface IPositionChanger
-    {
-        void ChangePosition(Vector3 motionForce);
-    }
-}
