--- Assets/Scripts/Features/Services/Transformation/IPositionChanger.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 23d02599c6b04b8abe8b31c5b9784c80
-timeCreated: 1674223999
