+++ Assets/Tests/Utilities/Features/Services/Transformation/Create.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using WizardSpells.Data.Dynamic;
+using WizardSpells.Features.Services.Force.MotionForce;
 using WizardSpells.Features.Services.Transformation;
 using WizardSpells.Tests.Utilities.Common.Extensions;
 
@@ -15,7 +16,8 @@ namespace WizardSpells.Tests.Utilities.Features.Services.Transformation
                 .MakeGrounded();
 
         public static PositionChanger PositionChanger(CharacterController characterController = default,
-            ICharacterData data = default) =>
-            new(characterController ? characterController : CharacterController(), data.SubstituteIfDefault());
+            ICharacterData data = default, IMotionForceProvider motionForceProvider = default) =>
+            new(characterController ? characterController : CharacterController(),
+                data.SubstituteIfDefault(), motionForceProvider.SubstituteIfDefault());
     }
 }
